+++ opentelemetry-appender-log/src/lib.rs
@@ -91,7 +91,6 @@
 //!
 //! This library provides the following Cargo features:
 //!
-//! - `spec_unstable_logs_enabled`: Allow users to control the log level.
 //! - `with-serde`: Support complex values as attributes without stringifying them.
 //!
 //! [Logs Bridge API]: https://opentelemetry.io/docs/specs/otel/logs/bridge-api/
@@ -135,14 +134,11 @@ where
     L: Logger + Send + Sync,
 {
     fn enabled(&self, _metadata: &Metadata) -> bool {
-        #[cfg(feature = "spec_unstable_logs_enabled")]
-        return self.logger.event_enabled(
+        self.logger.event_enabled(
             severity_of_level(_metadata.level()),
             _metadata.target(),
             None,
-        );
-        #[cfg(not(feature = "spec_unstable_logs_enabled"))]
-        true
+        )
     }
 
     fn log(&self, record: &Record) {
@@ -791,9 +787,6 @@ mod tests {
         // As a result of using `with_simple_exporter` while building the logger provider,
         // the processor used is a `SimpleLogProcessor` which has an implementation of `event_enabled`
         // that always returns true.
-        #[cfg(feature = "spec_unstable_logs_enabled")]
-        assert!(otel_log_appender.enabled(&log::Metadata::builder().build()));
-        #[cfg(not(feature = "spec_unstable_logs_enabled"))]
         assert!(otel_log_appender.enabled(&log::Metadata::builder().build()));
     }