+++ opentelemetry/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 
+- **Breaking** Stabilized `logger.enabled` by removing the `spec_unstable_logs_enabled` feature flag. The `event_enabled` method on `Logger` trait is now always available.
 - Add `reserve` method to `opentelemetry::propagation::Injector` to hint at the number of elements that will be added to avoid multiple resize operations of the underlying data structure. Has an empty default implementation.
 
 ## v0.31.0